requirements.txt
asgiref==3.8.1
astroid==3.2.4
attrs==24.2.0
coverage==7.6.1
dill==0.3.8
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
drf-spectacular==0.27.2
execnet==2.1.1
flake8==7.1.1
flake8-report==0.1.0a1
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
mccabe==0.7.0
packaging==24.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pycodestyle==2.12.1
pyflakes==3.2.0
PyJWT==2.9.0
pylint==3.2.6
pymongo==4.8.0
pytest==8.3.2
pytest-cov==5.0.0
pytest-django==4.8.0
pytest-xdist==3.6.1
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.0
sqlparse==0.5.1
tomlkit==0.13.0
uritemplate==4.1.1
whitenoise==6.7.0
